🔵 extract `AddQueryParameters`
