+++ cmd/stdcov-cli/main.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"strings"
 
 	"gitlab.com/multi/stdcov-api-test/cmd/stdcov-cli/client"
 )
@@ -36,11 +35,7 @@ func Run() int {
 		return 1
 	}
 
-	q := req.URL.Query()
-	for k, v := range query.params {
-		q.Add(k, v)
-	}
-	req.URL.RawQuery = q.Encode()
+	AddQueryParameters(query, req)
 
 	report, err := ExecuteTestSuite(c, req)
 	if err != nil {
@@ -54,33 +49,4 @@ func Run() int {
 		return 1
 	}
 	return 0
-}
-
-// Query implements flag.Value interface to store query parameters
-type Query struct {
-	params map[string]string
-}
-
-// String implements flag.Value.String
-func (qp *Query) String() string {
-	str := ""
-	for k, v := range qp.params {
-		str += fmt.Sprintf("--%s:%s ", k, v)
-	}
-	return str
-}
-
-// Set implements flag.Value.Set
-func (qp *Query) Set(s string) error {
-	parts := strings.SplitN(s, "=", 2)
-	key := parts[0]
-	value := ""
-	if len(parts) > 1 {
-		value = parts[1]
-	}
-	if qp.params == nil {
-		qp.params = make(map[string]string)
-	}
-	qp.params[key] = value
-	return nil
 }
